+++ src/components/v2/V2Project/V2ProjectSettings/pages/V2PaymentAddressSettingsPage.tsx
@@ -1,5 +1,8 @@
-import React from 'react'
+import { PaymentAddressSection } from 'components/Project/ProjectToolsDrawer/PaymentAddressSection'
+import { useDeployProjectPayerTx } from 'hooks/v2/transactor/DeployProjectPayerTx'
 
 export function V2PaymentAddressSettingsPage() {
-  return <div>V2PaymentAddressSettingsPage</div>
+  return (
+    <PaymentAddressSection useDeployProjectPayerTx={useDeployProjectPayerTx} />
+  )
 }
