+++ src/components/v2/V2Project/V2ProjectToolsDrawer/V1TokenMigrationSetupSection/V1TokenMigrationSetupSection.tsx
@@ -20,10 +20,7 @@ export function V1TokenMigrationSetupSection() {
   const hasV1ProjectId = Boolean(v1Project?.toNumber() ?? 0 > 0)
 
   return (
-    <section>
-      <h3>
-        <Trans>V1 token migration</Trans>
-      </h3>
+    <>
       <p>
         <Trans>
           Allow your V1 project token holders to swap their tokens for your V2
@@ -72,6 +69,6 @@ export function V1TokenMigrationSetupSection() {
         onOk={() => setMigrationModalVisible(false)}
         onCancel={() => setMigrationModalVisible(false)}
       />
-    </section>
+    </>
   )
 }
